requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
Flask==2.2.3
python-dotenv==1.0.0
requests==2.25.1
selenium==4.8.2
webdriver-manager==3.8.5
